# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: AnimusForge API
  description: |
    ## Adaptive AI Persona Framework with Ethical Governance

    AnimusForge is an enterprise-grade framework for creating, managing, and evolving
    AI personas with built-in ethical governance, kill-switch mechanisms, and
    comprehensive memory systems.

    ### Key Features

    - **Persona Management**: Create, update, activate, and evolve AI personas with
      configurable traits, cognitive patterns, and behavioral constraints.
    - **LLM Gateway**: Multi-provider LLM access with automatic failover, load balancing,
      and cost optimization across OpenAI, Anthropic, DeepSeek, and GLM.
    - **Kill-Switch**: Emergency stop and recovery mechanisms for immediate persona
      deactivation with audit trail and recovery workflows.
    - **Memory System**: Vector-based semantic memory and graph-based knowledge storage
      for persistent persona context and learning.
    - **Ethics Engine**: Real-time content evaluation, audit logging, and compliance
      reporting across multiple ethical categories.
    - **System Monitoring**: Health checks, Prometheus metrics, and configuration
      management endpoints.

    ### Authentication

    JWT authentication is supported. Include `Authorization: Bearer <token>` header
    for authenticated requests. Some endpoints may require elevated permissions.

    ### Rate Limiting

    All endpoints are rate-limited. Check response headers:
...
  version: 1.0.0
  contact:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT authentication token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
  schemas:
...
    UUID:
      type: string
      format: uuid
      example: "550e8400-e29b-41d4-a716-446655440000"
    Timestamp:
      type: string
      format: date-time
      example: "2024-03-06T12:00:00Z"
...
    KillSwitchTrigger:
      type: string
      enum:
        - manual
        - safety_violation
        - ethics_violation
        - behavior_anomaly
        - resource_limit
        - security_breach
        - cascade
        - scheduled
    KillSwitchSeverity:
      type: string
      enum:
        - low
        - medium
        - high
        - critical
        - emergency
    KillSwitchTriggerRequest:
      type: object
      required:
        - persona_id
        - trigger_type
        - reason
      properties:
        persona_id:
          $ref: "#/components/schemas/UUID"
        trigger_type:
          $ref: "#/components/schemas/KillSwitchTrigger"
        severity:
          $ref: "#/components/schemas/KillSwitchSeverity"
        reason:
          type: string
          minLength: 1
          maxLength: 2000
        immediate:
          type: boolean
          default: true
        cascade:
          type: boolean
          default: false
    KillSwitchTriggerResponse:
      type: object
      properties:
        incident_id:
          $ref: "#/components/schemas/UUID"
        persona_id:
          $ref: "#/components/schemas/UUID"
        status:
          type: string
          enum: [triggered, active, recovering, recovered, failed]
        triggered_at:
          $ref: "#/components/schemas/Timestamp"
        severity:
          $ref: "#/components/schemas/KillSwitchSeverity"
        estimated_recovery_time:
          type: integer
          description: Seconds until recovery available

    # ==================== Memory Schemas ====================
...
paths:
...
  /api/v1/killswitch/{persona_id}/trigger:
    post:
      tags:
        - Kill-Switch
      summary: Trigger kill-switch
      description: Trigger the kill-switch for a specific persona to immediately halt operations.
      operationId: triggerKillswitch
      parameters:
        - name: persona_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KillSwitchTriggerRequest"
      responses:
        "201":
          description: Kill-switch triggered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KillSwitchTriggerResponse"
        "400":
          description: Kill-switch already active
  /api/v1/killswitch/{persona_id}/status:
    get:
      tags:
        - Kill-Switch
      summary: Get kill-switch status
      description: Retrieve the current kill-switch status for a persona.
      operationId: getKillswitchStatus
      parameters:
        - name: persona_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: Kill-switch status
  /api/v1/killswitch/{persona_id}/recover:
    post:
      tags:
        - Kill-Switch
      summary: Recover from kill-switch
      description: Initiate recovery process to restore persona operations after kill-switch.
      operationId: recoverFromKillswitch
      parameters:
        - name: persona_id
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - strategy
              properties:
                strategy:
                  type: string
                  enum: [automatic, manual, incremental, checkpoint, full_reset, rollback]
                auto_validate:
                  type: boolean
                  default: true
      responses:
        "200":
          description: Recovery initiated successfully
        "400":
          description: No active kill-switch
  /api/v1/killswitch/incidents:
    get:
      tags:
        - Kill-Switch
      summary: List incidents
      description: Retrieve a paginated list of kill-switch incidents.
      operationId: listIncidents
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
        - name: persona_id
          in: query
          schema:
            $ref: "#/components/schemas/UUID"
        - name: severity
          in: query
          schema:
            $ref: "#/components/schemas/KillSwitchSeverity"
      responses:
        "200":
          description: List of incidents

  # ==================== Memory ====================
...